+++ prowler/lib/outputs/outputs.py
@@ -28,7 +28,7 @@ def stdout_report(finding, color, verbose, status, fix):
         details = finding.check_metadata.CheckID
     if finding.check_metadata.Provider == "iac":
         details = finding.check_metadata.CheckID
-    if finding.check_metadata.Provider == "oci":
+    if finding.check_metadata.Provider == "oraclecloud":
         details = finding.region
 
     if (verbose or fix) and (not status or finding.status in status):